添加系统管理页面 修复教师和管理员控制器返回错误

+++ ExamOnline/src/henu/dao/impl/TeacherDaoImpl.java
@@ -73,7 +73,7 @@ public class TeacherDaoImpl implements TeacherDao {
 	 */
 	private long getCount() throws SQLException {
 		String sql = "SELECT COUNT(*) FROM teacher;";
-		return qr.query(sql, new ScalarHandler<>());
+		return (long) qr.query(sql, new ScalarHandler<>());
 
 	}
 }
